+++ CHANGELOG.md
@@ -44,6 +44,7 @@ It also supports the new "added" timestamp for songs in the mpd database.
 - Feat: Filling the jukebox queue is executed in worker thread
 - Feat: Rework structs to improve thread safety
 - Feat: Improve random number generation
+- Feat: Add option to show audioformat in the footer #1184
 - Upd: Improve number conversions and enforcement of limits
 - Upd: Manual add random songs or albums should ignore the jukebox constraints
 - Upd: libmympdclient 1.0.25